+++ src/style.css
@@ -27,14 +27,18 @@ body {
 input::placeholder {
   font-style: italic;
 }
-input{
+
+input {
   border: none;
 }
+
 /* style.css */
-/* Add some basic styles to the todo items and the edit/delete buttons */
-/* style.css */
+
 /* Add some basic styles to the todo items and the edit/delete buttons */
 
+/* style.css */
+
+/* Add some basic styles to the todo items and the edit/delete buttons */
 
 .todo-item label {
   flex: 1;
@@ -54,7 +58,9 @@ input{
 .todo-item:hover .fa-trash {
   display: inline;
 }
-.todo-item:hover .fa-ellipsis-vertical{
-display: none;
+
+.todo-item:hover .fa-ellipsis-vertical {
+  display: none;
 }
+
 /* ... Other styles ... */
